moved all storybook files out of stories folder to being next to their respective component/page
