+++ apps/ui-sharethrift/src/components/layouts/home/Homepage.stories.tsx
@@ -1,7 +1,7 @@
 import type { Meta, StoryFn } from "@storybook/react";
-import { HomeRoutes } from "../index.tsx";
-import { ListingsPageContainerGetListingsDocument } from "../../../../generated.tsx";
-import { withMockApolloClient, withMockRouter } from "../../../../test-utils/storybook-decorators.tsx";
+import { HomeRoutes } from "./index.tsx";
+import { ListingsPageContainerGetListingsDocument } from "../../../generated.tsx";
+import { withMockApolloClient, withMockRouter } from "../../../test-utils/storybook-decorators.tsx";
 import { expect, within } from 'storybook/test';
 
 const meta: Meta<typeof HomeRoutes> = {
