+++ apps/ui-sharethrift/src/components/layouts/home/account/admin-dashboard/components/admin-listings-table/admin-listings-table.status-filter.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react';
 import { expect, within, fn, userEvent } from 'storybook/test';
-import { StatusFilter } from '../admin-listings-table.status-filter.tsx';
+import { StatusFilter } from './admin-listings-table.status-filter.tsx';
 
 const meta: Meta<typeof StatusFilter> = {
 	title: 'Admin/ListingsTable/StatusFilter',
