+++ apps/ui-sharethrift/src/components/layouts/home/account/admin-dashboard/components/admin-listings-table/admin-listings-table.title-filter.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react';
 import { expect, within, fn, userEvent } from 'storybook/test';
-import { TitleFilter } from '../admin-listings-table.title-filter.tsx';
+import { TitleFilter } from './admin-listings-table.title-filter.tsx';
 
 const meta: Meta<typeof TitleFilter> = {
 	title: 'Admin/ListingsTable/TitleFilter',
