+++ apps/ui-sharethrift/src/components/layouts/home/section-layout.stories.tsx
@@ -4,8 +4,8 @@ import { MemoryRouter } from 'react-router-dom';
 import { ApolloClient, InMemoryCache } from '@apollo/client';
 import { ApolloProvider } from '@apollo/client/react';
 import { MockLink } from '@apollo/client/testing';
-import { MockAuthWrapper } from '../../../../test-utils/storybook-decorators.tsx';
-import { SectionLayout } from '../section-layout.tsx';
+import { MockAuthWrapper } from '../../../test-utils/storybook-decorators.tsx';
+import { SectionLayout } from './section-layout.tsx';
 
 // Mock Apollo Client with MockLink
 const mockApolloClient = new ApolloClient({
