+++ apps/ui-sharethrift/src/components/layouts/signup/section-layout.stories.tsx
@@ -1,27 +1,32 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import { expect, within, userEvent } from 'storybook/test';
+import { expect, userEvent, within } from 'storybook/test';
+import { MemoryRouter } from 'react-router-dom';
+import { ApolloClient, InMemoryCache } from '@apollo/client';
+import { ApolloProvider } from '@apollo/client/react';
+import { MockLink } from '@apollo/client/testing';
+import { MockAuthWrapper } from '../../../test-utils/storybook-decorators.tsx';
 import { SectionLayout } from './section-layout.tsx';
-import { MemoryRouter, Routes, Route } from 'react-router-dom';
-import { MockAuthWrapper, withMockApolloClient } from '../../../test-utils/storybook-decorators.tsx';
+
+const mockApolloClient = new ApolloClient({
+	link: new MockLink([]),
+	cache: new InMemoryCache(),
+});
 
 const meta: Meta<typeof SectionLayout> = {
-	title: 'Layouts/SectionLayout',
+	title: 'Layouts/SignupLayout',
 	component: SectionLayout,
 	parameters: {
 		layout: 'fullscreen',
 	},
 	decorators: [
-		withMockApolloClient,
 		(Story) => (
-			<MockAuthWrapper>
-				<MemoryRouter initialEntries={['/signup']}>
-					<Routes>
-						<Route path="/signup" element={<Story />}>
-							<Route index element={<div>Signup Content</div>} />
-						</Route>
-					</Routes>
-				</MemoryRouter>
-			</MockAuthWrapper>
+			<ApolloProvider client={mockApolloClient}>
+				<MockAuthWrapper>
+					<MemoryRouter initialEntries={['/signup']}>
+						<Story />
+					</MemoryRouter>
+				</MockAuthWrapper>
+			</ApolloProvider>
 		),
 	],
 };
@@ -31,20 +36,73 @@ type Story = StoryObj<typeof SectionLayout>;
 
 export const Default: Story = {
 	play: async ({ canvasElement }) => {
-		const canvas = within(canvasElement);
-		await expect(canvasElement).toBeTruthy();
-		await expect(canvas.getByText('Signup Content')).toBeInTheDocument();
+		const header = canvasElement.querySelector('header');
+		await expect(header).toBeInTheDocument();
+
+		const main = canvasElement.querySelector('main');
+		await expect(main).toBeInTheDocument();
+
+		const footer = canvasElement.querySelector('footer');
+		await expect(footer).toBeInTheDocument();
 	},
 };
 
-export const ClickHeaderButtons: Story = {
+export const WithEnvironmentHandling: Story = {
+	play: async ({ canvasElement }) => {
+		const header = canvasElement.querySelector('header');
+		await expect(header).toBeInTheDocument();
+	},
+};
+export const ClickLogoutButton: Story = {
 	play: async ({ canvasElement }) => {
 		const canvas = within(canvasElement);
-		
-		const loginButtons = canvas.queryAllByRole('button', { name: /Log In/i });
-		const firstButton = loginButtons[0];
-		if (firstButton) {
-			await userEvent.click(firstButton);
+		await expect(canvasElement).toBeTruthy();
+		const logoutButton = canvas.queryByText(/Sign Out|Logout/i);
+		if (logoutButton) {
+			await userEvent.click(logoutButton);
+		}
+	},
+};
+
+export const ProductionModeLogin: Story = {
+	parameters: {
+		env: {
+			MODE: 'production',
+		},
+	},
+	play: async ({ canvasElement }) => {
+		const canvas = within(canvasElement);
+		await expect(canvasElement).toBeTruthy();
+		const loginButton = canvas.queryByText(/Login|Sign In/i);
+		if (loginButton) {
+			await userEvent.click(loginButton);
+		}
+	},
+};
+
+export const ProductionModeAdminLogin: Story = {
+	parameters: {
+		env: {
+			MODE: 'production',
+		},
+	},
+	play: async ({ canvasElement }) => {
+		const canvas = within(canvasElement);
+		await expect(canvasElement).toBeTruthy();
+		const adminButton = canvas.queryByText(/Admin/i);
+		if (adminButton) {
+			await userEvent.click(adminButton);
+		}
+	},
+};
+
+export const ClickSignUp: Story = {
+	play: async ({ canvasElement }) => {
+		const canvas = within(canvasElement);
+		await expect(canvasElement).toBeTruthy();
+		const signupButton = canvas.queryByText(/Sign Up/i);
+		if (signupButton) {
+			await userEvent.click(signupButton);
 		}
 	},
 };
@@ -52,10 +110,10 @@ export const ClickHeaderButtons: Story = {
 export const ClickCreateListing: Story = {
 	play: async ({ canvasElement }) => {
 		const canvas = within(canvasElement);
-		
-		const shareButton = canvas.queryByRole('button', { name: /Share/i });
-		if (shareButton) {
-			await userEvent.click(shareButton);
+		await expect(canvasElement).toBeTruthy();
+		const createButton = canvas.queryByText(/Create Listing|Post/i);
+		if (createButton) {
+			await userEvent.click(createButton);
 		}
 	},
 };
