+++ apps/ui-sharethrift/src/components/shared/login-selection.stories.tsx
@@ -1,8 +1,8 @@
 import type { Meta, StoryObj } from '@storybook/react';
 import { expect, within, userEvent } from 'storybook/test';
 import { MemoryRouter } from 'react-router-dom';
-import { MockAuthWrapper } from '../../../test-utils/storybook-decorators.tsx';
-import { LoginSelection } from '../login-selection.tsx';
+import { MockAuthWrapper } from '../../test-utils/storybook-decorators.tsx';
+import { LoginSelection } from './login-selection.tsx';
 
 const meta: Meta<typeof LoginSelection> = {
 	title: 'Shared/LoginSelection',
